# swagger/output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Social Network for Music | API",
    "description": "A RESTful API that works as the core of SNM (Social Network for Music).",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "JWT": {
      "type": "apiKey",
      "in": "header",
      "name": "authorization",
      "description": "Enter the JWT with the 'Bearer ' prefix (e.g. 'Bearer [token]')."
    }
  },
  "paths": {
...
    "/api/spotify/tracks/": {
      "get": {
        "tags": [
          "spotify"
        ],
        "summary": "Allows to search spotify's tracks",
        "description": "",
        "parameters": [
          {
            "name": "q",
            "description": "Searches tracks that match the following query",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Track"
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/spotify/artists/": {
      "get": {
        "tags": [
          "spotify"
        ],
        "summary": "Allows to search spotify's artists",
        "description": "",
        "parameters": [
          {
            "name": "q",
            "description": "Searches artists that match the following query",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Artist"
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/spotify/genres/": {
      "get": {
        "tags": [
          "spotify"
        ],
        "summary": "Retrieves all spotify's genres",
        "description": "",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "genres": {
                  "type": "array",
                  "example": [
                    "acoustic",
                    "afrobeat",
                    "alt-rock",
                    "..."
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/spotify/recommendations/": {
      "get": {
        "tags": [
          "spotify"
        ],
        "summary": "Returns the current user's recommendations",
        "description": "",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Track"
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/spotify/tracks/{id}/": {
      "get": {
        "tags": [
          "spotify"
        ],
        "summary": "Retrieves a spotify's track",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "A track's ID"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Track"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/spotify/artists/{id}/": {
      "get": {
        "tags": [
          "spotify"
        ],
        "summary": "Retrieves a spotify's artist",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "An artist's ID"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Artist"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    }
  },
  "definitions": {
...
    "Track": {
      "type": "object",
      "properties": {
        "album": {
          "type": "object",
          "properties": {
            "album_type": {
              "type": "string",
              "example": "album"
            },
            "artists": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "external_urls": {
                    "type": "object",
                    "properties": {
                      "spotify": {
                        "type": "string",
                        "example": "https://open.spotify.com/artist/67hb7towEyKvt5Z8Bx306c"
                      }
                    }
                  },
                  "href": {
                    "type": "string",
                    "example": "https://api.spotify.com/v1/artists/67hb7towEyKvt5Z8Bx306c"
                  },
                  "id": {
                    "type": "string",
                    "example": "67hb7towEyKvt5Z8Bx306c"
                  },
                  "name": {
                    "type": "string",
                    "example": "Empire Of The Sun"
                  },
                  "type": {
                    "type": "string",
                    "example": "artist"
                  },
                  "uri": {
                    "type": "string",
                    "example": "spotify:artist:67hb7towEyKvt5Z8Bx306c"
                  }
                }
              }
            },
            "available_markets": {
              "type": "array",
              "example": [
                "AR",
                "AU",
                "AT",
                "BE",
                "BO",
                "BR",
                "BG",
                "CA",
                "CL",
                "CO",
                "CR",
                "CY",
                "CZ",
                "DK",
                "DO",
                "DE",
                "EC",
                "EE",
                "SV",
                "FI",
                "FR",
                "GR",
                "GT",
                "HN",
                "HK",
                "HU",
                "IS",
                "IE",
                "IT",
                "LV",
                "LT",
                "LU",
                "MY",
                "MT",
                "MX",
                "NL",
                "NZ",
                "NI",
                "NO",
                "PA",
                "PY",
                "PE",
                "PH",
                "PL",
                "PT",
                "SG",
                "SK",
                "ES",
                "SE",
                "CH",
                "TW",
                "TR",
                "UY",
                "US",
                "GB",
                "AD",
                "LI",
                "MC",
                "ID",
                "JP",
                "TH",
                "VN",
                "RO",
                "IL",
                "ZA",
                "SA",
                "AE",
                "BH",
                "QA",
                "OM",
                "KW",
                "EG",
                "MA",
                "DZ",
                "TN",
                "LB",
                "JO",
                "PS",
                "IN",
                "BY",
                "KZ",
                "MD",
                "UA",
                "AL",
                "BA",
                "HR",
                "ME",
                "MK",
                "RS",
                "SI",
                "KR",
                "BD",
                "PK",
                "LK",
                "GH",
                "KE",
                "NG",
                "TZ",
                "UG",
                "AG",
                "AM",
                "BS",
                "BB",
                "BZ",
                "BT",
                "BW",
                "BF",
                "CV",
                "CW",
                "DM",
                "FJ",
                "GM",
                "GE",
                "GD",
                "GW",
                "GY",
                "HT",
                "JM",
                "KI",
                "LS",
                "LR",
                "MW",
                "MV",
                "ML",
                "MH",
                "FM",
                "NA",
                "NR",
                "NE",
                "PW",
                "PG",
                "WS",
                "SM",
                "ST",
                "SN",
                "SC",
                "SL",
                "SB",
                "KN",
                "LC",
                "VC",
                "SR",
                "TL",
                "TO",
                "TT",
                "TV",
                "VU",
                "AZ",
                "BN",
                "BI",
                "KH",
                "CM",
                "TD",
                "KM",
                "GQ",
                "SZ",
                "GA",
                "GN",
                "KG",
                "LA",
                "MO",
                "MR",
                "MN",
                "NP",
                "RW",
                "TG",
                "UZ",
                "ZW",
                "BJ",
                "MG",
                "MU",
                "MZ",
                "AO",
                "CI",
                "DJ",
                "ZM",
                "CD",
                "CG",
                "IQ",
                "LY",
                "TJ",
                "VE",
                "ET",
                "XK"
              ],
              "items": {
                "type": "string"
              }
            },
            "external_urls": {
              "type": "object",
              "properties": {
                "spotify": {
                  "type": "string",
                  "example": "https://open.spotify.com/album/5B6XfyHHYawyLkEvNvhSPh"
                }
              }
            },
            "href": {
              "type": "string",
              "example": "https://api.spotify.com/v1/albums/5B6XfyHHYawyLkEvNvhSPh"
            },
            "id": {
              "type": "string",
              "example": "5B6XfyHHYawyLkEvNvhSPh"
            },
            "images": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "height": {
                    "type": "number",
                    "example": 64
                  },
                  "url": {
                    "type": "string",
                    "example": "https://i.scdn.co/image/ab67616d00004851f3aa0e6ca22a382007f61e4d"
                  },
                  "width": {
                    "type": "number",
                    "example": 64
                  }
                }
              }
            },
            "name": {
              "type": "string",
              "example": "Walking On A Dream (10th Anniversary Edition)"
            },
            "release_date": {
              "type": "string",
              "example": "2008-10-03"
            },
            "release_date_precision": {
              "type": "string",
              "example": "day"
            },
            "total_tracks": {
              "type": "number",
              "example": 16
            },
            "type": {
              "type": "string",
              "example": "album"
            },
            "uri": {
              "type": "string",
              "example": "spotify:album:5B6XfyHHYawyLkEvNvhSPh"
            }
          }
        },
        "artists": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "external_urls": {
                "type": "object",
                "properties": {
                  "spotify": {
                    "type": "string",
                    "example": "https://open.spotify.com/artist/67hb7towEyKvt5Z8Bx306c"
                  }
                }
              },
              "href": {
                "type": "string",
                "example": "https://api.spotify.com/v1/artists/67hb7towEyKvt5Z8Bx306c"
              },
              "id": {
                "type": "string",
                "example": "67hb7towEyKvt5Z8Bx306c"
              },
              "name": {
                "type": "string",
                "example": "Empire Of The Sun"
              },
              "type": {
                "type": "string",
                "example": "artist"
              },
              "uri": {
                "type": "string",
                "example": "spotify:artist:67hb7towEyKvt5Z8Bx306c"
              }
            }
          }
        },
        "available_markets": {
          "type": "array",
          "example": [
            "AR",
            "AU",
            "AT",
            "BE",
            "BO",
            "BR",
            "BG",
            "CA",
            "CL",
            "CO",
            "CR",
            "CY",
            "CZ",
            "DK",
            "DO",
            "DE",
            "EC",
            "EE",
            "SV",
            "FI",
            "FR",
            "GR",
            "GT",
            "HN",
            "HK",
            "HU",
            "IS",
            "IE",
            "IT",
            "LV",
            "LT",
            "LU",
            "MY",
            "MT",
            "MX",
            "NL",
            "NZ",
            "NI",
            "NO",
            "PA",
            "PY",
            "PE",
            "PH",
            "PL",
            "PT",
            "SG",
            "SK",
            "ES",
            "SE",
            "CH",
            "TW",
            "TR",
            "UY",
            "US",
            "GB",
            "AD",
            "LI",
            "MC",
            "ID",
            "JP",
            "TH",
            "VN",
            "RO",
            "IL",
            "ZA",
            "SA",
            "AE",
            "BH",
            "QA",
            "OM",
            "KW",
            "EG",
            "MA",
            "DZ",
            "TN",
            "LB",
            "JO",
            "PS",
            "IN",
            "BY",
            "KZ",
            "MD",
            "UA",
            "AL",
            "BA",
            "HR",
            "ME",
            "MK",
            "RS",
            "SI",
            "KR",
            "BD",
            "PK",
            "LK",
            "GH",
            "KE",
            "NG",
            "TZ",
            "UG",
            "AG",
            "AM",
            "BS",
            "BB",
            "BZ",
            "BT",
            "BW",
            "BF",
            "CV",
            "CW",
            "DM",
            "FJ",
            "GM",
            "GE",
            "GD",
            "GW",
            "GY",
            "HT",
            "JM",
            "KI",
            "LS",
            "LR",
            "MW",
            "MV",
            "ML",
            "MH",
            "FM",
            "NA",
            "NR",
            "NE",
            "PW",
            "PG",
            "WS",
            "SM",
            "ST",
            "SN",
            "SC",
            "SL",
            "SB",
            "KN",
            "LC",
            "VC",
            "SR",
            "TL",
            "TO",
            "TT",
            "TV",
            "VU",
            "AZ",
            "BN",
            "BI",
            "KH",
            "CM",
            "TD",
            "KM",
            "GQ",
            "SZ",
            "GA",
            "GN",
            "KG",
            "LA",
            "MO",
            "MR",
            "MN",
            "NP",
            "RW",
            "TG",
            "UZ",
            "ZW",
            "BJ",
            "MG",
            "MU",
            "MZ",
            "AO",
            "CI",
            "DJ",
            "ZM",
            "CD",
            "CG",
            "IQ",
            "LY",
            "TJ",
            "VE",
            "ET",
            "XK"
          ],
          "items": {
            "type": "string"
          }
        },
        "disc_number": {
          "type": "number",
          "example": 1
        },
        "duration_ms": {
          "type": "number",
          "example": 198440
        },
        "explicit": {
          "type": "boolean",
          "example": false
        },
        "external_ids": {
          "type": "object",
          "properties": {
            "isrc": {
              "type": "string",
              "example": "AUEI10800039"
            }
          }
        },
        "external_urls": {
          "type": "object",
          "properties": {
            "spotify": {
              "type": "string",
              "example": "https://open.spotify.com/track/5r5cp9IpziiIsR6b93vcnQ"
            }
          }
        },
        "href": {
          "type": "string",
          "example": "https://api.spotify.com/v1/tracks/5r5cp9IpziiIsR6b93vcnQ"
        },
        "id": {
          "type": "string",
          "example": "5r5cp9IpziiIsR6b93vcnQ"
        },
        "is_local": {
          "type": "boolean",
          "example": false
        },
        "name": {
          "type": "string",
          "example": "Walking On A Dream"
        },
        "popularity": {
          "type": "number",
          "example": 82
        },
        "preview_url": {},
        "track_number": {
          "type": "number",
          "example": 2
        },
        "type": {
          "type": "string",
          "example": "track"
        },
        "uri": {
          "type": "string",
          "example": "spotify:track:5r5cp9IpziiIsR6b93vcnQ"
        }
      }
    },
    "Artist": {
      "type": "object",
      "properties": {
        "external_urls": {
          "type": "object",
          "properties": {
            "spotify": {
              "type": "string",
              "example": "https://open.spotify.com/artist/67hb7towEyKvt5Z8Bx306c"
            }
          }
        },
        "followers": {
          "type": "object",
          "properties": {
            "href": {},
            "total": {
              "type": "number",
              "example": 1410412
            }
          }
        },
        "genres": {
          "type": "array",
          "example": [
            "australian dance",
            "australian electropop",
            "dance rock",
            "indietronica",
            "neo-synthpop"
          ],
          "items": {
            "type": "string"
          }
        },
        "href": {
          "type": "string",
          "example": "https://api.spotify.com/v1/artists/67hb7towEyKvt5Z8Bx306c"
        },
        "id": {
          "type": "string",
          "example": "67hb7towEyKvt5Z8Bx306c"
        },
        "images": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "height": {
                "type": "number",
                "example": 160
              },
              "url": {
                "type": "string",
                "example": "https://i.scdn.co/image/ab6761610000f178e51c443e9e8c1952c80357b4"
              },
              "width": {
                "type": "number",
                "example": 160
              }
            }
          }
        },
        "name": {
          "type": "string",
          "example": "Empire Of The Sun"
        },
        "popularity": {
          "type": "number",
          "example": 75
        },
        "type": {
          "type": "string",
          "example": "artist"
        },
        "uri": {
          "type": "string",
          "example": "spotify:artist:67hb7towEyKvt5Z8Bx306c"
        }
      }
    }
  }
}